pyproject.toml
[tool.poetry]
name = "2022-server"
version = "0.1.0"
description = ""
authors = ["aminhp93 <minhpn.org.ec@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.7"
django-cors-headers = "^3.7.0"
djangorestframework = "^3.12.4"
psycopg2 = "^2.8.6"
django-model-utils = "^4.1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
